core: move NUMA nodes to a dedicated special level
They have their own custom depth HWLOC_TYPE_DEPTH_NUMANODE (-7 just like IO/misc)
We can still iterate over them without any problem.

For now they are still in the main tree but that will change soon.

Given that we always have a NUMA level and its depth is constant,
we can simplify several places in backends and utils.

+++ utils/lstopo/lstopo-text.c
@@ -208,11 +208,11 @@ static void output_distances(struct lstopo_output *loutput)
   if (!err) {
     for(j=0; j<nr; j++) {
       const char *kindmeans = (dist[j]->kind & HWLOC_DISTANCES_KIND_MEANS_LATENCY) ? "latency" : (dist[j]->kind & HWLOC_DISTANCES_KIND_MEANS_BANDWIDTH) ? "bandwidth" : "distance";
-      fprintf(output, "Relative %s matrix (kind %lu) between %u %ss (depth %u) by %s indexes:\n",
+      fprintf(output, "Relative %s matrix (kind %lu) between %u %ss (depth %d) by %s indexes:\n",
 	      kindmeans, dist[j]->kind,
 	      dist[j]->nbobjs,
 	      hwloc_type_name(dist[j]->objs[0]->type),
-	      dist[j]->objs[0]->depth,
+	      (int) dist[j]->objs[0]->depth,
 	      logical ? "logical" : "physical");
       hwloc_utils_print_distance_matrix(output, dist[j]->nbobjs, dist[j]->objs, dist[j]->values, logical);
       hwloc_distances_release(topology, dist[j]);
